bugfixes, new stuff
-minor bugfixes, changes to some status, new (unimplemented) maps

+++ src/dev/zt/UpliftedVFFV/inventory/equipables/ClockwiseCage.java
@@ -25,7 +25,7 @@ public class ClockwiseCage extends Item{
 	}
 	
 	public status[] getEnchantment(Schmuck s) {
-		enchantment[0] = new BonusStatBuff(14,(int)(.20*(1+s.getEquipPow())), 50);
+		enchantment[0] = new BonusStatBuff(14,.20*(1+s.getEquipPow()), 50);
 		enchantment[1] = new BonusStatBuff(31,1, 50);
 		return enchantment;
 	}
